refactor: extract empty offer helpers
