+++ packages/midnight-sdk/src/signatures/TreeUtils.ts
@@ -15,69 +15,12 @@ import {
 } from "../offers/index.js";
 import { Group } from "./Group.js";
 import { type GroupInput, GroupUtils } from "./GroupUtils.js";
+import {
+  EMPTY_OFFER_STRUCT,
+  isEmptyOfferStruct,
+} from "./offerStructInternal.js";
 import { encode as encodePayload } from "./Payload.js";
 import type { Tree } from "./Tree.js";
-
-const ZERO_ADDRESS = "0x0000000000000000000000000000000000000000" as const;
-const ZERO_BYTES32 =
-  "0x0000000000000000000000000000000000000000000000000000000000000000" as const;
-
-const EMPTY_OFFER_STRUCT: OfferStruct = {
-  market: {
-    chainId: 0n,
-    midnight: ZERO_ADDRESS,
-    loanToken: ZERO_ADDRESS,
-    collateralParams: [],
-    maturity: 0n,
-    rcfThreshold: 0n,
-    enterGate: ZERO_ADDRESS,
-    liquidatorGate: ZERO_ADDRESS,
-  },
-  buy: false,
-  maker: ZERO_ADDRESS,
-  start: 0n,
-  expiry: 0n,
-  tick: 0n,
-  group: ZERO_BYTES32,
-  callback: ZERO_ADDRESS,
-  callbackData: "0x",
-  receiverIfMakerIsSeller: ZERO_ADDRESS,
-  ratifier: ZERO_ADDRESS,
-  reduceOnly: false,
-  maxUnits: 0n,
-  maxAssets: 0n,
-  continuousFeeCap: 0n,
-};
-
-const EMPTY_OFFER_DEFAULT_GROUP = OfferUtils.hashStruct(EMPTY_OFFER_STRUCT);
-
-function isEmptyOfferStruct(offer: OfferStruct): boolean {
-  return (
-    offer.market.chainId === 0n &&
-    offer.market.midnight === ZERO_ADDRESS &&
-    offer.market.loanToken === ZERO_ADDRESS &&
-    offer.market.collateralParams.length === 0 &&
-    offer.market.maturity === 0n &&
-    offer.market.rcfThreshold === 0n &&
-    offer.market.enterGate === ZERO_ADDRESS &&
-    offer.market.liquidatorGate === ZERO_ADDRESS &&
-    offer.buy === false &&
-    offer.maker === ZERO_ADDRESS &&
-    offer.start === 0n &&
-    offer.expiry === 0n &&
-    offer.tick === 0n &&
-    (offer.group === ZERO_BYTES32 ||
-      offer.group === EMPTY_OFFER_DEFAULT_GROUP) &&
-    offer.callback === ZERO_ADDRESS &&
-    offer.callbackData === "0x" &&
-    offer.receiverIfMakerIsSeller === ZERO_ADDRESS &&
-    offer.ratifier === ZERO_ADDRESS &&
-    offer.reduceOnly === false &&
-    offer.maxUnits === 0n &&
-    offer.maxAssets === 0n &&
-    offer.continuousFeeCap === 0n
-  );
-}
 
 function isPowerOfTwo(value: number): boolean {
   return value > 0 && (value & (value - 1)) === 0;
@@ -106,7 +49,7 @@ function assertLeafOffers(
 ): void {
   const seen = new Set<Hash>();
   for (const [index, offer] of offers.entries()) {
-    if (isEmptyOfferStruct(offer)) continue;
+    if (isEmptyOfferStruct(offer, { allowDefaultGroup: true })) continue;
 
     const leafHash = leafHashes[index]!;
     if (seen.has(leafHash)) {
